This PR uses ethnum instead of bnum for the arithmetic operations, except for addmod and mulmod because those are not yet supported (see https://github.com/nlordell/ethnum-rs/issues/10).
In the Rust implementation of ethnum u128 is used as the base type which is faster than u64 as observed in #959.

ethnum also has a llvm-intrinsics feature which uses the llvm 256 bit integer type. However, even with linker-plugin-lto I did not get it to inline those functions and without inlining this is slower than the pure Rust implementation.

```
                   │ 2024-12-06T12:40#6196d42#20-main/evmrs#performance │ 2024-12-06T14:20#b22d4e4#20/evmrs#performance │
                   │                       sec/op                       │        sec/op          vs base                │
StaticOverhead/1/                                           1.753µ ± 0%             1.802µ ± 1%   +2.82% (p=0.000 n=20)
Inc/1/                                                      4.112µ ± 0%             3.723µ ± 0%   -9.45% (p=0.000 n=20)
Inc/10/                                                     4.110µ ± 0%             3.710µ ± 0%   -9.72% (p=0.000 n=20)
Fib/1/                                                      3.506µ ± 0%             3.353µ ± 1%   -4.36% (p=0.000 n=20)
Fib/5/                                                      14.25µ ± 0%             12.48µ ± 0%  -12.39% (p=0.000 n=20)
Fib/10/                                                     137.2µ ± 0%             121.2µ ± 0%  -11.67% (p=0.000 n=20)
Fib/15/                                                     1.332m ± 0%             1.170m ± 0%  -12.15% (p=0.000 n=20)
Fib/20/                                                     14.45m ± 0%             12.58m ± 0%  -12.93% (p=0.000 n=20)
Sha3/1/                                                     2.053µ ± 0%             2.123µ ± 1%   +3.43% (p=0.000 n=20)
Sha3/10/                                                    3.469µ ± 0%             3.333µ ± 0%   -3.92% (p=0.000 n=20)
Sha3/100/                                                   17.00µ ± 1%             14.92µ ± 0%  -12.27% (p=0.000 n=20)
Sha3/1000/                                                  182.8µ ± 0%             151.4µ ± 0%  -17.17% (p=0.000 n=20)
Arith/1/                                                    4.632µ ± 1%             4.580µ ± 0%   -1.11% (p=0.001 n=20)
Arith/10/                                                  10.090µ ± 1%             8.610µ ± 1%  -14.67% (p=0.000 n=20)
Arith/100/                                                  64.13µ ± 0%             49.08µ ± 0%  -23.46% (p=0.000 n=20)
Arith/280/                                                  189.0µ ± 0%             165.7µ ± 0%  -12.34% (p=0.000 n=20)
Memory/1/                                                   6.608µ ± 1%             6.254µ ± 0%   -5.35% (p=0.000 n=20)
Memory/10/                                                  11.77µ ± 1%             10.92µ ± 1%   -7.21% (p=0.000 n=20)
Memory/100/                                                 64.01µ ± 0%             56.47µ ± 0%  -11.78% (p=0.000 n=20)
Memory/1000/                                                569.3µ ± 0%             508.1µ ± 0%  -10.75% (p=0.000 n=20)
Memory/10000/                                               5.377m ± 0%             4.749m ± 0%  -11.69% (p=0.000 n=20)
Analysis/jumpdest/                                          1.758µ ± 1%             1.825µ ± 1%   +3.81% (p=0.000 n=20)
Analysis/stop/                                              1.759µ ± 1%             1.812µ ± 1%   +3.07% (p=0.000 n=20)
Analysis/push1/                                             1.755µ ± 1%             1.818µ ± 0%   +3.56% (p=0.000 n=20)
Analysis/push32/                                            1.763µ ± 1%             1.817µ ± 1%   +3.06% (p=0.000 n=20)
geomean                                                     22.60µ                  20.87µ        -7.68%
```